Extract query class for Notification.
